Fixed some problems with values returned from lua scripts.
Specially when returning tables, arrays (which in lua are also tables), and arrays holding more arrays inside.

Also needed to add a kind of aliasing to the `eval` command as I couldn't use the `eval` keyword to define the
`eval` function for the eval command because `eval` is a reserved word in JS' strict mode.
So I decided to use some ad-hoc alias `evaluate` so, if the command is named `evaluate` then register it as `eval`.