Improve keyPrefix support
When you use `createConnectedClient` to share the same data with
different values for `keyPrefix`, each client needs to apply its
`keyPrefix` separately.

The previous implementation did not properly apply the new `keyPrefix`
provided as an option to `createConnectedClient`, resulting in all
clients having effectively the same `keyPrefix`.

This would break tests that are trying to use `keyPrefix` to avoid
key collisions.