Return table for HGETALL from custom Lua scripts
This commit fixes a bug that causes HGETALL commands in Lua
to fail with ioredis-mock.

Redis returns an array reply for HGETALL. ioredis and ioredis-mock
convert this response to a JS object.

When using custom Lua scripts with ioredis, however, HGETALL returns
and array-like table. With this change, HGETALL will also return an
array-like table from Lua with ioredis-mock.