fix: BLOCKING behaviour to not return immediately
It needs to wait until the timeElapsed is greater or eq the defined blocking time.
The test was enhanced to spot the issue and verify the fix worked. This
uncovered another issue with the way found events are evaluated to always return
even without any event received. This problem is also fixed by this commit.